requirements.txt
python-dotenv
pandas 
numpy
pymongo
certifi 
pymongo[srv]
scikit-learn
pyyaml
dill
xgboost
fastapi
uvicorn
apache-airflow
python-multipart
-e .